commit #3

+++ src/main/java/com/example/hulkstore/Service/ProductoService.java
@@ -6,23 +6,35 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductoService {
 
     @Autowired
-    ProductoRepository productoRepository;
+    private ProductoRepository productoRepository;
 
-    public List<Producto> getProductos(){
+    public List<Producto> obtenerProductos(){
         return productoRepository.findAll();
+    }
+
+    public Optional<Producto> obtenerProductoPorId(Long productoId){
+        return productoRepository.findById(productoId);
     }
 
     public void agregarProducto(Producto producto){
         productoRepository.save(producto);
     }
 
-    public void borrarProducto(Producto producto){
-        productoRepository.delete(producto);
+    public void borrarProducto(Long productoId){
+        Optional<Producto> productoOptional = productoRepository.findById(productoId);
+
+        if(productoOptional.isPresent()){
+            Producto producto = productoOptional.get();
+            productoRepository.delete(producto);
+        } else {
+            System.out.println("No se encuentra el producto");
+        }
     }
 
     public void actualizarProducto(Producto producto){
